requirements.txt
peewee==3.17.0
fastapi==0.108.0
python-dotenv==0.21.0
httpx==0.25.2
uvicorn==0.20.0
peewee-migrate==1.6.5
psycopg2-binary==2.9.9
# psycopg2==2.9.5
pydantic==2.5.3
sentry-sdk==1.39.1
redis==4.4.2
# fastapi-utils==0.2.1
flower==2.0.1
pyjwt==2.8.0
python-multipart==0.0.6
python-jose==3.3.0
ipython==8.11.0
passlib==1.7.4
bcrypt==3.2.0
twilio==8.11.0
